app/requirements.txt
Django==2.2.5
psycopg2-binary==2.8.3
djangorestframework==3.11.0
django-filter==2.2.0
django-debug-toolbar==2.0
drf-yasg==1.17.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
